feat: add initial admin layout with updated routing
